Reset system errno before negotiate
